Additional test transforms to illustrate wrong transformation
This commit actually belongs before commit that fixes the transformation.

The new test transforms are part of the "from" transforms. In Unity,
the test is run by manually first duplicating and then transforming
(translating, rotating and shrinking) the "from" transforms. But the
test transforms are not used to create the transformation and when
applying the resulting transformation to them, the mapping of the
"from" to the "to" transforms fails more reliably and with greater
distance between them than can be seen when applying the transformation to
the transforms that were used to create the that transformation.